chore(TS): Migrated filters module in store
